+++ Hook/Hook/Utils/Stream.swift
@@ -48,28 +48,3 @@ final class MutableStream<T: Equatable>: ReadOnlyStream<T>, MutableStreamType {
         relay.accept(value)
     }
 }
-
-extension MutableStream where T == [Tag: [Bookmark]] {
-    
-    func update(with bookmark: Bookmark) {
-        var dictionary = value
-        var tags: [Tag]
-        
-        if let bookmarkTags = bookmark.tags {
-            tags = [Tag(name: TagName.all)] + bookmarkTags.map({ Tag(name: $0.name) })
-        } else {
-            tags = [Tag(name: TagName.all)]
-        }
-        
-        tags.forEach {
-            if var bookmarks = dictionary[$0] {
-                bookmarks.insert(bookmark, at: 0)
-                dictionary.updateValue(bookmarks, forKey: $0)
-            } else {
-                dictionary.updateValue([bookmark], forKey: $0)
-            }
-        }
-        
-        update(with: dictionary)
-    }
-}
